requirements.txt
flask
requests
msal
msgraph-sdk
azure-identity
sentence_transformers 
ctransformers
langchain-community
pdfplumber
pytesseract
streamlit
llama-cpp-python
chromadb
psycopg2-binary
langchain~=0.2.12
openai==0.28.0
sqlalchemy
psycopg2-binary             # For PostgreSQL connection
faiss-cpu==1.8.0           # For FAISS vector store
python-dotenv   
streamlit-msal
PyPDF2
python-docx
Pillow
openpyxl
tiktoken
matplotlib
wordcloud
seaborn
streamlit_option_menu
PyPDF
streamlit-extras
altair
watchdog
packaging
nltk
pydantic
atlassian-python-api
pytesseract
pdf2image
python-poppler
docx2txt
xlrd